added UserDTO, mappers and ManyToMany relationship user-specialty

+++ src/main/java/com/p2f4/med_office/entity/UserRole.java
@@ -14,7 +14,7 @@ public class UserRole {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_role", nullable = false)
-    private Integer idRole;
+    private Short idRole;
 
     @Column(name = "name", nullable = false)
     private String name;
